requirements.txt
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
openpyxl==3.0.7
pandas==1.3.2
pdf2image==1.16.0
Pillow==8.3.1
pycodestyle==2.7.0
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
